chore(meta): move fillCmpMetaFromConstructor to utils
